crematorium immune + banishment remind me to update the guidebook later

+++ Content.Server/Morgue/CrematoriumSystem.cs
@@ -45,6 +45,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+using Content.Goobstation.Shared.CrematorImmune;
 using Content.Server.Ghost;
 using Content.Server.Morgue.Components;
 using Content.Server.Storage.Components;
@@ -180,6 +181,9 @@ public sealed class CrematoriumSystem : EntitySystem
             for (var i = storage.Contents.ContainedEntities.Count - 1; i >= 0; i--)
             {
                 var item = storage.Contents.ContainedEntities[i];
+                if (HasComp<CrematoriumImmuneComponent>(item))
+                    continue;
+
                 _containers.Remove(item, storage.Contents);
                 EntityManager.DeleteEntity(item);
             }
